requirements.txt
fastapi==0.115.0
pydantic==2.9.0
pydantic-core==2.23.2   # <-- match Pydantic 2.9.0
uvicorn==0.30.1
gunicorn==21.2.0
sqlalchemy==2.0.35
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.2.0
openpyxl==3.1.2
